phase2-gradient-core/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
pydantic==2.11.7
langgraph==1.0.0
PyYAML==6.0.2
jsonschema==4.25.1
gradient-adk==0.2.11
python-multipart==0.0.20
fpdf2==2.8.4
pypdf==5.9.0
pytest==8.4.2
pytest-asyncio==1.1.0
